requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
anthropic==0.39.0
python-dotenv==1.0.1
python-multipart==0.0.12
httpx==0.27.2
psycopg2-binary==2.9.9
apscheduler==3.10.4
